Fix 3PAR driver handling of existing VLUNs
Changes the 3PAR FC and iSCSI drivers to check for existing
VLUNs by taking into consideration both the volume and the host
when querying the backend.

Also pulled out some common code from the FC and iSCSI driver
to reduce duplicate code relating to the finding of existing
VLUNs.

This is a backport to stable/Juno of bug/1475064 and Kilo review;
https://review.openstack.org/#/c/227088/